Consolidate some Inline* classes
